+++ sdk/servicebus/azure-servicebus/azure/servicebus/__init__.py
@@ -36,6 +36,7 @@ TransportType = constants.TransportType
 __all__ = [
     "ServiceBusMessage",
     "ServiceBusMessageBatch",
+    "ServiceBusMessageState",
     "ServiceBusReceivedMessage",
     "NEXT_AVAILABLE_SESSION",
     "ServiceBusSubQueue",